Unmarshal Links type to 'links' annotated struct fields
These changes add the ability to unmarshal links members to
links-annotated struct fields.

The specification for links members can be found here:
https://jsonapi.org/format/#document-links

Note that this does not change the existing marshaling behavior of
links (the Linkable interface methods), though a fully fleshed out
implementation probably should - it's awkward that now interface methods
are used to marshal but struct fields for unmarshaling.

In other words, an implementation of links marshaling similar to the
implementation proposed in PR 58 of the canonical google/jsonapi
repository would be most appropriate to pair with these unmarshaling
changes. The actual implementation that was accepted is PR 57, utilizing
those Linkable/Metable interfaces.

From 57:

> After looking at both implementations, I think I'd like to vote for
this one because it feels flexible, and doesn't add any additional
fields to our structs. In general I don't think LINK and META objects
really have much to do with the resource objects themselves, they are
really more related to the context in which the object is used. This
approach keeps structs cleaner.

This is subjectively wrong, and assumes that this package is used in the
context of a server and not a client. In a client, links members can
contain information that is vital to the functionality of the resource
itself, so we add that here.