Add an ignore annotation
Previously a struct could never contain a public field that could be
ignored by the JSONAPI marshalling/unmarshalling. This made it difficult
to provide a struct that only serialized part of the fields. Additionally
the JSON (and BSON) marshallers support the '-' as an ignore annotation.

This commit adds the "-" annotation and adds tests and docs updates for this
new annotation type.